core-time: explicitly cast ret to size_t
Avoid implicity casting of int ret to size_t by explicitly casting it.

+++ core-time.c
@@ -76,7 +76,7 @@ static inline void stress_format_time(
 			ret = snprintf(*ptr, *len, "%lu %s%s, ", val, units,
 				(val > 1) ? "s" : "");
 		if (ret > 0) {
-			*len -= ret;
+			*len -= (size_t)ret;
 			*ptr += ret;
 		}
 	}
